Add integration tests around template recompilation
This is a stab at recreating a testing framework around the
supervisor, starting with some tests around restart behavior as
triggered by hook and configuration recompilation.

For each test case, we use the `hab-sup` binary (along with
`hab-launch`; see below) to start a testing Habitat package (using
just the minimal amount of files, not a full-blown .hart
package). Then, we connect to the supervisor with a Butterfly client
to inject configuration changes. Finally, we inspect the contents of
files within the package's service directory to verify the desired
behaviors.

(For now, we'll also disable these tests on Windows platforms; we'll
basically need to duplicate the fixtures we have now in Powershell
versions.)

Because `hab-sup` no longer runs independently, but requires the
`hab-launch` binary, the Makefile has been tweaked to ensure this
binary exists before the supervisor tests are run. Unfortunately,
Cargo will not build binaries for dependent crates, so we can't just
depend on the launcher crate. We can't modify the `build.rs` script to
run an "inner Cargo run" to build `hab-launch` because Cargo obtains
an exclusive lock on a lock file in the output directory, ensuring
that only one Cargo run can exist a a time. Modifying the Makefile
appears to be the most straightforward way to get this working.

(Alternatively, we could basically end up duplicating the launcher
server and running that in a separate thread in our tests, but that
seems like a lot of unnecessary extra work; in any event, using the
hab-launch binary means our supervisor integration tests more
accurately reflect real-world use cases. They also serve as a kind of
smoke test for the launcher itself.)

(Incidentally, a few of the integration tests in this commit will
fail; they cover the bug that will be fixed in the following commit.)

Signed-off-by: Christopher Maier <cmaier@chef.io>